+++ index.js
@@ -310,7 +310,7 @@ if (!nativeBinding) {
   throw new Error(`Failed to load native binding`)
 }
 
-const { JsClient, JsPostgresProvider, JsSqliteProvider, JsRuntime, activityTraceLog, orchestrationTraceLog, activityIsCancelled, activityGetClient, activityTag, orchestrationSetCustomStatus, orchestrationResetCustomStatus, orchestrationGetCustomStatus, initTracing } = nativeBinding
+const { JsClient, JsPostgresProvider, JsSqliteProvider, JsRuntime, activityTraceLog, orchestrationTraceLog, activityIsCancelled, activityGetClient, activityTag, orchestrationSetCustomStatus, orchestrationResetCustomStatus, orchestrationGetCustomStatus, orchestrationSetValue, orchestrationGetValue, orchestrationClearValue, orchestrationClearAllValues, initTracing } = nativeBinding
 
 module.exports.JsClient = JsClient
 module.exports.JsPostgresProvider = JsPostgresProvider
@@ -324,4 +324,8 @@ module.exports.activityTag = activityTag
 module.exports.orchestrationSetCustomStatus = orchestrationSetCustomStatus
 module.exports.orchestrationResetCustomStatus = orchestrationResetCustomStatus
 module.exports.orchestrationGetCustomStatus = orchestrationGetCustomStatus
+module.exports.orchestrationSetValue = orchestrationSetValue
+module.exports.orchestrationGetValue = orchestrationGetValue
+module.exports.orchestrationClearValue = orchestrationClearValue
+module.exports.orchestrationClearAllValues = orchestrationClearAllValues
 module.exports.initTracing = initTracing